Stop logging character dummies not being able to equip accessories (#94635)
## About The Pull Request

Tin, noticed this spam in my log from a character preview that wasn't
wearing a uniform, and they had an accessory in their loadout.

It should not be doing this stuff in the char prefs menu when it's be
possible to preview uniformless characters.

## Why It's Good For The Game

Stops senseless world.log spam

## Changelog

Not player-facing